sort notes by reversed rank
