+++ src/routes/Board/Board.tsx
@@ -10,11 +10,8 @@ import {toast} from "react-toastify";
 import {Actions} from "store/action";
 import _ from "underscore";
 import {Outlet} from "react-router-dom";
-import {DragDropContext, DragUpdate, DropResult} from "react-beautiful-dnd";
-import {useDispatch} from "react-redux";
 
 export const Board = () => {
-  const dispatch = useDispatch();
   useEffect(
     () => () => {
       toast.clearWaitingQueue();
@@ -62,25 +59,6 @@ export const Board = () => {
     return <LoadingScreen />;
   }
 
-  const onDragEnd = (result: DropResult) => {
-    const {destination, source, combine, draggableId} = result;
-    if (combine) {
-      dispatch(Actions.editNote(draggableId, {position: {column: combine.droppableId, stack: combine.draggableId, rank: 0}}));
-    }
-    if (!destination || (destination.droppableId === source.droppableId && destination.index === source.index)) {
-      return;
-    }
-    dispatch(Actions.editNote(draggableId, {position: {column: destination.droppableId, stack: undefined, rank: destination.index}}));
-  };
-
-  const onDragUpdate = (initial: DragUpdate) => {
-    const {destination, source, draggableId} = initial;
-    if (!destination || (destination.droppableId === source.droppableId && destination.index === source.index)) {
-      return;
-    }
-    dispatch(Actions.editNote(draggableId, {position: {column: destination.droppableId, stack: undefined, rank: destination.index}}));
-  };
-
   if (state.board.status === "ready") {
     return (
       <>
@@ -92,23 +70,21 @@ export const Board = () => {
         )}
         <InfoBar />
         <Outlet />
-        <DragDropContext onDragEnd={onDragEnd} onDragUpdate={onDragUpdate}>
-          <BoardComponent currentUserIsModerator={currentUserIsModerator} moderating={state.view.moderating}>
-            {state.columns
-              .filter((column) => column.visible || (currentUserIsModerator && state.participants?.self.showHiddenColumns))
-              .map((column, columnIndex) => (
-                <Column
-                  tabIndex={TabIndex.Column + columnIndex * 17}
-                  key={column.id}
-                  id={column.id}
-                  index={column.index}
-                  name={column.name}
-                  visible={column.visible}
-                  color={column.color}
-                />
-              ))}
-          </BoardComponent>
-        </DragDropContext>
+        <BoardComponent currentUserIsModerator={currentUserIsModerator} moderating={state.view.moderating}>
+          {state.columns
+            .filter((column) => column.visible || (currentUserIsModerator && state.participants?.self.showHiddenColumns))
+            .map((column, columnIndex) => (
+              <Column
+                tabIndex={TabIndex.Column + columnIndex * 17}
+                key={column.id}
+                id={column.id}
+                index={column.index}
+                name={column.name}
+                visible={column.visible}
+                color={column.color}
+              />
+            ))}
+        </BoardComponent>
       </>
     );
   }
